Fix for Cafcass/Cafcass Cymru flag
