+++ src/main/java/uk/gov/hmcts/reform/prl/services/C100IssueCaseService.java
@@ -6,6 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.prl.enums.YesOrNo;
+import uk.gov.hmcts.reform.prl.models.court.CourtVenue;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.rpa.mappers.C100JsonMapper;
 import uk.gov.hmcts.reform.prl.services.document.DocumentGenService;
@@ -14,8 +15,10 @@ import uk.gov.hmcts.reform.prl.utils.CaseUtils;
 
 import java.time.LocalDate;
 import java.util.Map;
+import java.util.Optional;
 
 import static java.util.Objects.requireNonNull;
+import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.COURT_NAME_FIELD;
 
 @Service
 @Slf4j
@@ -37,35 +40,28 @@ public class C100IssueCaseService {
             requireNonNull(caseData);
             sendgridService.sendEmail(c100JsonMapper.map(caseData));
         }
-        caseData = caseData.toBuilder().issueDate(LocalDate.now()).build();
         Map<String, Object> caseDataUpdated = callbackRequest.getCaseDetails().getData();
 
-        /*if (null != caseData.getCourtList() && null != caseData.getCourtList().getValue()) {
-            String[] idEmail = caseData.getCourtList().getValue().getCode().split(":");
-            String baseLocationId = Arrays.stream(idEmail).toArray()[0].toString();
+        if (null != caseData.getCourtList() && null != caseData.getCourtList().getValue()) {
+            String baseLocationId = caseData.getCourtList().getValue().getCode();
             Optional<CourtVenue> courtVenue = locationRefDataService.getCourtDetailsFromEpimmsId(
                 baseLocationId,
                 authorisation
             );
-            String caseTypeOfApplication = CaseUtils.getCaseTypeOfApplication(caseData);
             caseDataUpdated.putAll(CaseUtils.getCourtDetails(courtVenue, baseLocationId));
-            caseDataUpdated.putAll(CaseUtils.getCourtEmail(idEmail, caseTypeOfApplication));
-            caseData.setCourtName(caseDataUpdated.get("courtName").toString());
-        }*/
-        caseData.setIssueDate();
+            caseData = caseData.toBuilder().issueDate(LocalDate.now())
+                .courtName(caseDataUpdated.containsKey(COURT_NAME_FIELD) ? caseDataUpdated.get(COURT_NAME_FIELD).toString() : null)
+                .build();
+        }
+
         // Generate All Docs and set to casedataupdated.
         caseDataUpdated.putAll(documentGenService.generateDocuments(authorisation, caseData));
 
         // Refreshing the page in the same event. Hence no external event call needed.
         // Getting the tab fields and add it to the casedetails..
         Map<String, Object> allTabsFields = allTabsService.getAllTabsFields(caseData);
-        caseDataUpdated.put("issueDate", LocalDate.now());
         caseDataUpdated.putAll(allTabsFields);
-        try {
-            caseWorkerEmailService.sendEmailToCourtAdmin(callbackRequest.getCaseDetails().toBuilder().data(caseDataUpdated).build());
-        } catch (Exception ex) {
-            log.error("Email notification could not be sent", ex);
-        }
+        caseDataUpdated.put("issueDate", caseData.getIssueDate());
         return caseDataUpdated;
     }
