+++ src/main/java/uk/gov/hmcts/reform/prl/services/LocationRefDataService.java
@@ -64,7 +64,7 @@ public class LocationRefDataService {
         return DynamicListElement.builder().code(key).label(value).build();
     }
 
-    public String getCourtDetailsFromEpimmsId(String baseLocationId, String authToken) {
+    public  Optional<CourtVenue> getCourtDetailsFromEpimmsId(String baseLocationId, String authToken) {
         CourtDetails courtDetails = locationRefDataApi.getCourtDetailsByService(authToken,
                                                                                 authTokenGenerator.generate(),
                                                                                 SERVICE_ID);
@@ -74,8 +74,6 @@ public class LocationRefDataService {
             .filter(location -> baseLocationId.equalsIgnoreCase(location.getCourtEpimmsId()))
             .findFirst();
 
-        return courtVenue.map(venue -> venue.getCourtEpimmsId() + "-" + venue.getRegionId()
-            + "-" + venue.getCourtName() + "-" + venue.getPostcode() + "-" + venue.getRegion()
-            + "-" + venue.getSiteName()).orElse("");
+        return courtVenue;
     }
 }
