+++ src/test/java/uk/gov/hmcts/reform/prl/controllers/FL401SubmitApplicationControllerTest.java
@@ -17,38 +17,32 @@ import uk.gov.hmcts.reform.idam.client.models.UserDetails;
 import uk.gov.hmcts.reform.prl.clients.CourtFinderApi;
 import uk.gov.hmcts.reform.prl.constants.PrlAppsConstants;
 import uk.gov.hmcts.reform.prl.enums.FL401OrderTypeEnum;
-import uk.gov.hmcts.reform.prl.enums.FamilyHomeEnum;
 import uk.gov.hmcts.reform.prl.enums.Gender;
 import uk.gov.hmcts.reform.prl.enums.LanguagePreference;
-import uk.gov.hmcts.reform.prl.enums.LivingSituationEnum;
-import uk.gov.hmcts.reform.prl.enums.PeopleLivingAtThisAddressEnum;
 import uk.gov.hmcts.reform.prl.enums.State;
-import uk.gov.hmcts.reform.prl.enums.YesNoBothEnum;
 import uk.gov.hmcts.reform.prl.enums.YesNoDontKnow;
 import uk.gov.hmcts.reform.prl.enums.YesOrNo;
 import uk.gov.hmcts.reform.prl.models.Address;
-import uk.gov.hmcts.reform.prl.models.Element;
 import uk.gov.hmcts.reform.prl.models.Organisation;
 import uk.gov.hmcts.reform.prl.models.common.dynamic.DynamicList;
 import uk.gov.hmcts.reform.prl.models.common.dynamic.DynamicListElement;
-import uk.gov.hmcts.reform.prl.models.complextypes.ChildrenLiveAtAddress;
-import uk.gov.hmcts.reform.prl.models.complextypes.Home;
 import uk.gov.hmcts.reform.prl.models.complextypes.LinkToCA;
 import uk.gov.hmcts.reform.prl.models.complextypes.PartyDetails;
 import uk.gov.hmcts.reform.prl.models.complextypes.TypeOfApplicationOrders;
 import uk.gov.hmcts.reform.prl.models.court.Court;
 import uk.gov.hmcts.reform.prl.models.court.CourtAddress;
 import uk.gov.hmcts.reform.prl.models.court.CourtEmailAddress;
+import uk.gov.hmcts.reform.prl.models.court.CourtVenue;
 import uk.gov.hmcts.reform.prl.models.court.ServiceArea;
 import uk.gov.hmcts.reform.prl.models.documents.Document;
 import uk.gov.hmcts.reform.prl.models.dto.GeneratedDocumentInfo;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CallbackResponse;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseDetails;
-import uk.gov.hmcts.reform.prl.models.language.DocumentLanguage;
 import uk.gov.hmcts.reform.prl.services.CaseWorkerEmailService;
 import uk.gov.hmcts.reform.prl.services.ConfidentialityTabService;
 import uk.gov.hmcts.reform.prl.services.CourtFinderService;
+import uk.gov.hmcts.reform.prl.services.FL401SubmitApplicationService;
 import uk.gov.hmcts.reform.prl.services.LocationRefDataService;
 import uk.gov.hmcts.reform.prl.services.SolicitorEmailService;
 import uk.gov.hmcts.reform.prl.services.UserService;
@@ -66,16 +60,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertNotNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.COURT_EMAIL_ADDRESS_FIELD;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.COURT_NAME_FIELD;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.DOCUMENT_FIELD_C8;
 import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.DOCUMENT_FIELD_C8_WELSH;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.DOCUMENT_FIELD_FINAL;
 import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.DOCUMENT_FIELD_FINAL_WELSH;
 
 @PropertySource(value = "classpath:application.yaml")
@@ -134,6 +124,9 @@ public class FL401SubmitApplicationControllerTest {
     private LocationRefDataService locationRefDataService;
 
     @Mock
+    private FL401SubmitApplicationService fl401SubmitApplicationService;
+
+    @Mock
     private CourtFinderApi courtFinderApi;
 
     public static final String authToken = "Bearer TestAuthToken";
@@ -188,7 +181,12 @@ public class FL401SubmitApplicationControllerTest {
         when(courtFinderApi.getCourtDetails(Mockito.anyString())).thenReturn(horshamCourt);
         when(courtFinderService.getEmailAddress(horshamCourt)).thenReturn(Optional.of(courtEmailAddress));
         when(locationRefDataService.getCourtDetailsFromEpimmsId(Mockito.anyString(), Mockito.anyString()))
-            .thenReturn("test-1-test-test-test-test");
+            .thenReturn(Optional.of(CourtVenue.builder()
+                                        .courtName("test")
+                                        .regionId("1")
+                                        .siteName("test")
+                                        .region("test")
+                                        .build()));
     }
 
     @Test
@@ -292,784 +290,6 @@ public class FL401SubmitApplicationControllerTest {
     }
 
     @Test
-    public void testCourtNameAndEmailAddressReturnedWhileFamilyEmailAddressReturned() throws Exception {
-
-        generatedDocumentInfo = GeneratedDocumentInfo.builder()
-            .url("TestUrl")
-            .binaryUrl("binaryUrl")
-            .hashToken("testHashToken")
-            .build();
-
-        PartyDetails applicant = PartyDetails.builder().representativeFirstName("Abc")
-            .representativeLastName("Xyz")
-            .gender(Gender.male)
-            .email("abc@xyz.com")
-            .canYouProvideEmailAddress(YesOrNo.Yes)
-            .phoneNumber("1234567890")
-            .isEmailAddressConfidential(YesOrNo.Yes)
-            .isPhoneNumberConfidential(YesOrNo.Yes)
-            .solicitorOrg(Organisation.builder().organisationID("ABC").organisationName("XYZ").build())
-            .solicitorAddress(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .doTheyHaveLegalRepresentation(YesNoDontKnow.yes)
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .draftOrderDoc(Document.builder()
-                               .documentUrl(generatedDocumentInfo.getUrl())
-                               .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                               .documentHash(generatedDocumentInfo.getHashToken())
-                               .documentFileName("FL401-Final.docx")
-                               .build())
-            .applicantsFL401(applicant)
-            .caseTypeOfApplication(PrlAppsConstants.FL401_CASE_TYPE)
-            .state(State.AWAITING_FL401_SUBMISSION_TO_HMCTS)
-            .submitCountyCourtSelection(dynamicList)
-            .build();
-
-        CallbackResponse callbackResponse = CallbackResponse.builder()
-            .data(CaseData.builder()
-                      .draftOrderDoc(Document.builder()
-                                         .documentUrl(generatedDocumentInfo.getUrl())
-                                         .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                                         .documentHash(generatedDocumentInfo.getHashToken())
-                                         .documentFileName("FL401-Final.docx")
-                                         .build())
-                      .state(State.AWAITING_SUBMISSION_TO_HMCTS)
-                      .build())
-            .build();
-
-        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-
-
-        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
-            .CallbackRequest.builder()
-            .caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder()
-                             .id(123L)
-                             .data(stringObjectMap)
-                             .build())
-            .build();
-
-        when(documentGenService.generateDocuments(Mockito.anyString(), Mockito.any(CaseData.class))).thenReturn(
-            fl401DocsMap);
-
-        UserDetails userDetails = UserDetails.builder()
-            .forename("test")
-            .surname("test")
-            .build();
-        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
-
-        AboutToStartOrSubmitCallbackResponse response = fl401SubmitApplicationController
-            .fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-
-
-
-        assertTrue(response.getData().containsKey(COURT_EMAIL_ADDRESS_FIELD));
-        assertTrue(response.getData().containsKey(COURT_NAME_FIELD));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_C8));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_FINAL));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_C8_WELSH));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_FINAL_WELSH));
-    }
-
-    @Test
-    public void testC8Formgenerationbasedconconfidentiality() throws Exception {
-        generatedDocumentInfo = GeneratedDocumentInfo.builder()
-            .url("TestUrl")
-            .binaryUrl("binaryUrl")
-            .hashToken("testHashToken")
-            .build();
-
-        List<FL401OrderTypeEnum> orderList = new ArrayList<>();
-
-        orderList.add(FL401OrderTypeEnum.occupationOrder);
-
-        orders = TypeOfApplicationOrders.builder()
-            .orderType(orderList)
-            .build();
-
-        linkToCA = LinkToCA.builder()
-            .linkToCaApplication(YesOrNo.Yes)
-            .caApplicationNumber("123")
-            .build();
-
-        PartyDetails applicant = PartyDetails.builder()
-            .representativeFirstName("Abc")
-            .representativeLastName("Xyz")
-            .gender(Gender.male)
-            .email("abc@xyz.com")
-            .canYouProvideEmailAddress(YesOrNo.Yes)
-            .canYouProvidePhoneNumber(YesOrNo.Yes)
-            .phoneNumber("1234567890")
-            .isEmailAddressConfidential(YesOrNo.No)
-            .isAddressConfidential(YesOrNo.No)
-            .isPhoneNumberConfidential(YesOrNo.No)
-            .address(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .solicitorOrg(Organisation.builder().organisationID("ABC").organisationName("XYZ").build())
-            .solicitorAddress(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .doTheyHaveLegalRepresentation(YesNoDontKnow.yes)
-            .build();
-
-        ChildrenLiveAtAddress childrenLiveAtAddress = ChildrenLiveAtAddress.builder()
-            .keepChildrenInfoConfidential(YesOrNo.No)
-            .childFullName("child")
-            .childsAge("12")
-            .isRespondentResponsibleForChild(YesOrNo.Yes)
-            .build();
-
-        Home homefull = Home.builder()
-            .address(Address.builder().addressLine1("123").build())
-            .everLivedAtTheAddress(YesNoBothEnum.yesApplicant)
-            .doesApplicantHaveHomeRights(YesOrNo.No)
-            .doAnyChildrenLiveAtAddress(YesOrNo.Yes)
-            .children(List.of(Element.<ChildrenLiveAtAddress>builder().value(childrenLiveAtAddress).build()))
-            .isPropertyRented(YesOrNo.No)
-            .isThereMortgageOnProperty(YesOrNo.No)
-            .isPropertyAdapted(YesOrNo.No)
-            .peopleLivingAtThisAddress(List.of(PeopleLivingAtThisAddressEnum.applicant))
-            .familyHome(List.of(FamilyHomeEnum.payForRepairs))
-            .livingSituation(List.of(LivingSituationEnum.awayFromHome))
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .caseTypeOfApplication(PrlAppsConstants.FL401_CASE_TYPE)
-            .typeOfApplicationOrders(orders)
-            .typeOfApplicationLinkToCA(linkToCA)
-            .draftOrderDoc(Document.builder()
-                               .documentUrl(generatedDocumentInfo.getUrl())
-                               .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                               .documentHash(generatedDocumentInfo.getHashToken())
-                               .documentFileName("FL401-Final.docx")
-                               .build())
-            .applicantsFL401(applicant)
-            .home(homefull)
-            .submitCountyCourtSelection(dynamicList)
-            .state(State.AWAITING_FL401_SUBMISSION_TO_HMCTS)
-            .build();
-
-        CallbackResponse callbackResponse = CallbackResponse.builder()
-            .data(CaseData.builder()
-                      .draftOrderDoc(Document.builder()
-                                         .documentUrl(generatedDocumentInfo.getUrl())
-                                         .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                                         .documentHash(generatedDocumentInfo.getHashToken())
-                                         .documentFileName("FL401-Final.docx")
-                                         .build())
-                      .state(State.AWAITING_SUBMISSION_TO_HMCTS)
-                      .build())
-            .build();
-
-        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-        when(documentGenService.generateDocuments(Mockito.anyString(), any(CaseData.class)))
-            .thenReturn(stringObjectMap);
-
-        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-
-        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
-            .CallbackRequest.builder()
-            .caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder()
-                             .id(123L)
-                             .data(stringObjectMap)
-                             .build())
-            .build();
-
-        Court closestDomesticAbuseCourt = courtFinderService.getNearestFamilyCourt(
-            CaseUtils.getCaseData(callbackRequest.getCaseDetails(), objectMapper));
-        Optional<CourtEmailAddress> matchingEmailAddress = courtFinderService.getEmailAddress(closestDomesticAbuseCourt);
-
-        when(courtFinderService.getNearestFamilyCourt(CaseUtils.getCaseData(
-            callbackRequest.getCaseDetails(),
-            objectMapper
-        )))
-            .thenReturn(court);
-
-        UserDetails userDetails = UserDetails.builder()
-            .forename("test")
-            .surname("test")
-            .build();
-        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
-
-        DocumentLanguage documentLanguage = DocumentLanguage.builder().isGenEng(true).isGenWelsh(true).build();
-        fl401SubmitApplicationController.fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-
-        verify(documentGenService, times(1)).generateDocuments(
-            Mockito.anyString(),
-            any(CaseData.class)
-        );
-
-    }
-
-
-    @Test
-    public void testC8Formgenerationbasedconconfidentiality2() throws Exception {
-        generatedDocumentInfo = GeneratedDocumentInfo.builder()
-            .url("TestUrl")
-            .binaryUrl("binaryUrl")
-            .hashToken("testHashToken")
-            .build();
-
-        List<FL401OrderTypeEnum> orderList = new ArrayList<>();
-
-        orderList.add(FL401OrderTypeEnum.occupationOrder);
-
-        orders = TypeOfApplicationOrders.builder()
-            .orderType(orderList)
-            .build();
-
-        linkToCA = LinkToCA.builder()
-            .linkToCaApplication(YesOrNo.Yes)
-            .caApplicationNumber("123")
-            .build();
-
-        PartyDetails applicant = PartyDetails.builder()
-            .representativeFirstName("Abc")
-            .representativeLastName("Xyz")
-            .gender(Gender.male)
-            .email("abc@xyz.com")
-            .canYouProvideEmailAddress(YesOrNo.Yes)
-            .canYouProvidePhoneNumber(YesOrNo.Yes)
-            .phoneNumber("1234567890")
-            .isEmailAddressConfidential(YesOrNo.Yes)
-            .isAddressConfidential(YesOrNo.Yes)
-            .isPhoneNumberConfidential(YesOrNo.Yes)
-            .address(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .solicitorOrg(Organisation.builder().organisationID("ABC").organisationName("XYZ").build())
-            .solicitorAddress(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .doTheyHaveLegalRepresentation(YesNoDontKnow.yes)
-            .build();
-
-        ChildrenLiveAtAddress childrenLiveAtAddress = ChildrenLiveAtAddress.builder()
-            .keepChildrenInfoConfidential(YesOrNo.Yes)
-            .childFullName("child")
-            .childsAge("12")
-            .isRespondentResponsibleForChild(YesOrNo.Yes)
-            .build();
-
-        Home homefull = Home.builder()
-            .address(Address.builder().addressLine1("123").build())
-            .everLivedAtTheAddress(YesNoBothEnum.yesApplicant)
-            .doesApplicantHaveHomeRights(YesOrNo.No)
-            .doAnyChildrenLiveAtAddress(YesOrNo.Yes)
-            .children(List.of(Element.<ChildrenLiveAtAddress>builder().value(childrenLiveAtAddress).build()))
-            .isPropertyRented(YesOrNo.No)
-            .isThereMortgageOnProperty(YesOrNo.No)
-            .isPropertyAdapted(YesOrNo.No)
-            .peopleLivingAtThisAddress(List.of(PeopleLivingAtThisAddressEnum.applicant))
-            .familyHome(List.of(FamilyHomeEnum.payForRepairs))
-            .livingSituation(List.of(LivingSituationEnum.awayFromHome))
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .caseTypeOfApplication(PrlAppsConstants.FL401_CASE_TYPE)
-            .typeOfApplicationOrders(orders)
-            .typeOfApplicationLinkToCA(linkToCA)
-            .draftOrderDoc(Document.builder()
-                               .documentUrl(generatedDocumentInfo.getUrl())
-                               .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                               .documentHash(generatedDocumentInfo.getHashToken())
-                               .documentFileName("FL401-Final.docx")
-                               .build())
-            .applicantsFL401(applicant)
-            .home(homefull)
-            .submitCountyCourtSelection(dynamicList)
-            .state(State.AWAITING_FL401_SUBMISSION_TO_HMCTS)
-            .build();
-
-        CallbackResponse callbackResponse = CallbackResponse.builder()
-            .data(CaseData.builder()
-                      .draftOrderDoc(Document.builder()
-                                         .documentUrl(generatedDocumentInfo.getUrl())
-                                         .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                                         .documentHash(generatedDocumentInfo.getHashToken())
-                                         .documentFileName("FL401-Final.docx")
-                                         .build())
-                      .state(State.AWAITING_SUBMISSION_TO_HMCTS)
-                      .build())
-            .build();
-
-        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-
-        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-
-        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
-            .CallbackRequest.builder()
-            .caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder()
-                             .id(123L)
-                             .data(stringObjectMap)
-                             .build())
-            .build();
-
-        Court closestDomesticAbuseCourt = courtFinderService.getNearestFamilyCourt(
-            CaseUtils.getCaseData(callbackRequest.getCaseDetails(), objectMapper));
-        Optional<CourtEmailAddress> matchingEmailAddress = courtFinderService.getEmailAddress(closestDomesticAbuseCourt);
-
-        when(courtFinderService.getNearestFamilyCourt(CaseUtils.getCaseData(
-            callbackRequest.getCaseDetails(),
-            objectMapper
-        )))
-            .thenReturn(court);
-
-        UserDetails userDetails = UserDetails.builder()
-            .forename("test")
-            .surname("test")
-            .build();
-        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
-
-        DocumentLanguage documentLanguage = DocumentLanguage.builder().isGenEng(true).isGenWelsh(true).build();
-        fl401SubmitApplicationController.fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-        verify(documentGenService, times(1)).generateDocuments(
-            Mockito.anyString(),
-            any(CaseData.class)
-        );
-    }
-
-
-    @Test
-    public void testC8Formgenerationbasedconconfidentiality_withoutTypeofOrders() throws Exception {
-        generatedDocumentInfo = GeneratedDocumentInfo.builder()
-            .url("TestUrl")
-            .binaryUrl("binaryUrl")
-            .hashToken("testHashToken")
-            .build();
-
-
-        orders = TypeOfApplicationOrders.builder()
-            .orderType(null)
-            .build();
-
-        linkToCA = LinkToCA.builder()
-            .linkToCaApplication(YesOrNo.Yes)
-            .caApplicationNumber("123")
-            .build();
-
-        PartyDetails applicant = PartyDetails.builder()
-            .representativeFirstName("Abc")
-            .representativeLastName("Xyz")
-            .gender(Gender.male)
-            .email("abc@xyz.com")
-            .canYouProvideEmailAddress(YesOrNo.Yes)
-            .canYouProvidePhoneNumber(YesOrNo.Yes)
-            .phoneNumber("1234567890")
-            .isEmailAddressConfidential(YesOrNo.Yes)
-            .isAddressConfidential(YesOrNo.Yes)
-            .isPhoneNumberConfidential(YesOrNo.Yes)
-            .address(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .solicitorOrg(Organisation.builder().organisationID("ABC").organisationName("XYZ").build())
-            .solicitorAddress(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .doTheyHaveLegalRepresentation(YesNoDontKnow.yes)
-            .build();
-
-        ChildrenLiveAtAddress childrenLiveAtAddress = ChildrenLiveAtAddress.builder()
-            .keepChildrenInfoConfidential(YesOrNo.Yes)
-            .childFullName("child")
-            .childsAge("12")
-            .isRespondentResponsibleForChild(YesOrNo.Yes)
-            .build();
-
-        Home homefull = Home.builder()
-            .address(Address.builder().addressLine1("123").build())
-            .everLivedAtTheAddress(YesNoBothEnum.yesApplicant)
-            .doesApplicantHaveHomeRights(YesOrNo.No)
-            .doAnyChildrenLiveAtAddress(YesOrNo.Yes)
-            .children(List.of(Element.<ChildrenLiveAtAddress>builder().value(childrenLiveAtAddress).build()))
-            .isPropertyRented(YesOrNo.No)
-            .isThereMortgageOnProperty(YesOrNo.No)
-            .isPropertyAdapted(YesOrNo.No)
-            .peopleLivingAtThisAddress(List.of(PeopleLivingAtThisAddressEnum.applicant))
-            .familyHome(List.of(FamilyHomeEnum.payForRepairs))
-            .livingSituation(List.of(LivingSituationEnum.awayFromHome))
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .caseTypeOfApplication(PrlAppsConstants.FL401_CASE_TYPE)
-            .typeOfApplicationOrders(null)
-            .typeOfApplicationLinkToCA(linkToCA)
-            .draftOrderDoc(Document.builder()
-                               .documentUrl(generatedDocumentInfo.getUrl())
-                               .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                               .documentHash(generatedDocumentInfo.getHashToken())
-                               .documentFileName("FL401-Final.docx")
-                               .build())
-            .applicantsFL401(applicant)
-            .home(homefull)
-            .submitCountyCourtSelection(dynamicList)
-            .state(State.AWAITING_FL401_SUBMISSION_TO_HMCTS)
-            .build();
-
-        CallbackResponse callbackResponse = CallbackResponse.builder()
-            .data(CaseData.builder()
-                      .draftOrderDoc(Document.builder()
-                                         .documentUrl(generatedDocumentInfo.getUrl())
-                                         .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                                         .documentHash(generatedDocumentInfo.getHashToken())
-                                         .documentFileName("FL401-Final.docx")
-                                         .build())
-                      .state(State.AWAITING_SUBMISSION_TO_HMCTS)
-                      .build())
-            .build();
-
-        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-
-        when(documentGenService.generateDocuments(Mockito.anyString(), any(CaseData.class)))
-            .thenReturn(stringObjectMap);
-        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-
-        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
-            .CallbackRequest.builder()
-            .caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder()
-                             .id(123L)
-                             .data(stringObjectMap)
-                             .build())
-            .build();
-
-        Court closestDomesticAbuseCourt = courtFinderService.getNearestFamilyCourt(
-            CaseUtils.getCaseData(callbackRequest.getCaseDetails(), objectMapper));
-        Optional<CourtEmailAddress> matchingEmailAddress = courtFinderService.getEmailAddress(closestDomesticAbuseCourt);
-
-        when(courtFinderService.getNearestFamilyCourt(CaseUtils.getCaseData(
-            callbackRequest.getCaseDetails(),
-            objectMapper
-        )))
-            .thenReturn(court);
-        DocumentLanguage documentLanguage = DocumentLanguage.builder().isGenEng(true).isGenWelsh(true).build();
-
-        UserDetails userDetails = UserDetails.builder()
-            .forename("test")
-            .surname("test")
-            .build();
-        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
-
-        fl401SubmitApplicationController.fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-
-        verify(documentGenService, times(1)).generateDocuments(
-            Mockito.anyString(),
-            any(CaseData.class)
-        );
-    }
-
-    @Test
-    public void testC8Formgenerationbasedconconfidentiality_() throws Exception {
-        generatedDocumentInfo = GeneratedDocumentInfo.builder()
-            .url("TestUrl")
-            .binaryUrl("binaryUrl")
-            .hashToken("testHashToken")
-            .build();
-
-        List<FL401OrderTypeEnum> orderList = new ArrayList<>();
-
-        orderList.add(FL401OrderTypeEnum.occupationOrder);
-
-        orders = TypeOfApplicationOrders.builder()
-            .orderType(orderList)
-            .build();
-
-        linkToCA = LinkToCA.builder()
-            .linkToCaApplication(YesOrNo.Yes)
-            .caApplicationNumber("123")
-            .build();
-
-        PartyDetails applicant = PartyDetails.builder()
-            .representativeFirstName("Abc")
-            .representativeLastName("Xyz")
-            .gender(Gender.male)
-            .email("abc@xyz.com")
-            .canYouProvideEmailAddress(YesOrNo.Yes)
-            .canYouProvidePhoneNumber(YesOrNo.Yes)
-            .phoneNumber("1234567890")
-            .isEmailAddressConfidential(YesOrNo.Yes)
-            .isAddressConfidential(YesOrNo.Yes)
-            .isPhoneNumberConfidential(YesOrNo.Yes)
-            .address(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .solicitorOrg(Organisation.builder().organisationID("ABC").organisationName("XYZ").build())
-            .solicitorAddress(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .doTheyHaveLegalRepresentation(YesNoDontKnow.yes)
-            .build();
-
-        ChildrenLiveAtAddress childrenLiveAtAddress = ChildrenLiveAtAddress.builder()
-            .keepChildrenInfoConfidential(YesOrNo.Yes)
-            .childFullName("child")
-            .childsAge("12")
-            .isRespondentResponsibleForChild(YesOrNo.Yes)
-            .build();
-
-
-        CaseData caseData = CaseData.builder()
-            .caseTypeOfApplication(PrlAppsConstants.FL401_CASE_TYPE)
-            .typeOfApplicationOrders(orders)
-            .typeOfApplicationLinkToCA(linkToCA)
-            .draftOrderDoc(Document.builder()
-                               .documentUrl(generatedDocumentInfo.getUrl())
-                               .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                               .documentHash(generatedDocumentInfo.getHashToken())
-                               .documentFileName("FL401-Final.docx")
-                               .build())
-            .applicantsFL401(applicant)
-            .home(null)
-            .submitCountyCourtSelection(dynamicList)
-            .state(State.AWAITING_FL401_SUBMISSION_TO_HMCTS)
-            .build();
-
-        CallbackResponse callbackResponse = CallbackResponse.builder()
-            .data(CaseData.builder()
-                      .draftOrderDoc(Document.builder()
-                                         .documentUrl(generatedDocumentInfo.getUrl())
-                                         .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                                         .documentHash(generatedDocumentInfo.getHashToken())
-                                         .documentFileName("FL401-Final.docx")
-                                         .build())
-                      .state(State.AWAITING_SUBMISSION_TO_HMCTS)
-                      .build())
-            .build();
-
-        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-
-        when(documentGenService.generateDocuments(Mockito.anyString(), any(CaseData.class)))
-            .thenReturn(stringObjectMap);
-        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-
-        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
-            .CallbackRequest.builder()
-            .caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder()
-                             .id(123L)
-                             .data(stringObjectMap)
-                             .build())
-            .build();
-
-        Court closestDomesticAbuseCourt = courtFinderService.getNearestFamilyCourt(
-            CaseUtils.getCaseData(callbackRequest.getCaseDetails(), objectMapper));
-        Optional<CourtEmailAddress> matchingEmailAddress = courtFinderService.getEmailAddress(closestDomesticAbuseCourt);
-
-        when(courtFinderService.getNearestFamilyCourt(CaseUtils.getCaseData(
-            callbackRequest.getCaseDetails(),
-            objectMapper
-        )))
-            .thenReturn(court);
-
-        UserDetails userDetails = UserDetails.builder()
-            .forename("test")
-            .surname("test")
-            .build();
-        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
-
-        fl401SubmitApplicationController.fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-
-        verify(documentGenService, times(1)).generateDocuments(
-            Mockito.anyString(),
-            any(CaseData.class)
-        );
-
-    }
-
-    @Test
-    public void testCourtNameAndEmailAddressReturnedWhileFamilyEmailAddressReturned_WithNonMolestationOrder() throws Exception {
-
-        generatedDocumentInfo = GeneratedDocumentInfo.builder()
-            .url("TestUrl")
-            .binaryUrl("binaryUrl")
-            .hashToken("testHashToken")
-            .build();
-
-        List<FL401OrderTypeEnum> orderList = new ArrayList<>();
-
-        orderList.add(FL401OrderTypeEnum.nonMolestationOrder);
-
-        orders = TypeOfApplicationOrders.builder()
-            .orderType(orderList)
-            .build();
-
-        linkToCA = LinkToCA.builder()
-            .linkToCaApplication(YesOrNo.Yes)
-            .caApplicationNumber("123")
-            .build();
-
-        PartyDetails applicant = PartyDetails.builder()
-            .representativeFirstName("Abc")
-            .representativeLastName("Xyz")
-            .gender(Gender.male)
-            .email("abc@xyz.com")
-            .canYouProvideEmailAddress(YesOrNo.Yes)
-            .phoneNumber("1234567890")
-            .isEmailAddressConfidential(YesOrNo.Yes)
-            .isPhoneNumberConfidential(YesOrNo.Yes)
-            .solicitorOrg(Organisation.builder().organisationID("ABC").organisationName("XYZ").build())
-            .solicitorAddress(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .doTheyHaveLegalRepresentation(YesNoDontKnow.yes)
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .caseTypeOfApplication(PrlAppsConstants.FL401_CASE_TYPE)
-            .typeOfApplicationOrders(orders)
-            .typeOfApplicationLinkToCA(linkToCA)
-            .draftOrderDoc(Document.builder()
-                               .documentUrl(generatedDocumentInfo.getUrl())
-                               .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                               .documentHash(generatedDocumentInfo.getHashToken())
-                               .documentFileName("FL401-Final.docx")
-                               .build())
-            .applicantsFL401(applicant)
-            .state(State.AWAITING_FL401_SUBMISSION_TO_HMCTS)
-            .submitCountyCourtSelection(dynamicList)
-            .build();
-
-        CallbackResponse callbackResponse = CallbackResponse.builder()
-            .data(CaseData.builder()
-                      .draftOrderDoc(Document.builder()
-                                         .documentUrl(generatedDocumentInfo.getUrl())
-                                         .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                                         .documentHash(generatedDocumentInfo.getHashToken())
-                                         .documentFileName("FL401-Final.docx")
-                                         .build())
-                      .state(State.AWAITING_SUBMISSION_TO_HMCTS)
-                      .build())
-            .build();
-
-        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-
-        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
-            .CallbackRequest.builder()
-            .caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder()
-                             .id(123L)
-                             .data(stringObjectMap)
-                             .build())
-            .build();
-
-        Court closestDomesticAbuseCourt = courtFinderService.getNearestFamilyCourt(
-            CaseUtils.getCaseData(callbackRequest.getCaseDetails(), objectMapper));
-        Optional<CourtEmailAddress> matchingEmailAddress = courtFinderService.getEmailAddress(closestDomesticAbuseCourt);
-
-        when(courtFinderService.getNearestFamilyCourt(CaseUtils.getCaseData(
-            callbackRequest.getCaseDetails(),
-            objectMapper
-        )))
-            .thenReturn(court);
-        DocumentLanguage documentLanguage = DocumentLanguage.builder().isGenEng(true).isGenWelsh(true).build();
-
-        UserDetails userDetails = UserDetails.builder()
-            .forename("test")
-            .surname("test")
-            .build();
-        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
-
-        fl401SubmitApplicationController.fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-
-        when(documentGenService.generateDocuments(Mockito.anyString(), Mockito.any(CaseData.class))).thenReturn(
-            fl401DocsMap);
-        AboutToStartOrSubmitCallbackResponse response = fl401SubmitApplicationController
-            .fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_C8));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_FINAL));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_C8_WELSH));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_FINAL_WELSH));
-    }
-
-    @Test
-    public void testCourtNameAndEmailAddressReturnedWhileFamilyEmailAddressReturned_WithOccupationalOrder() throws Exception {
-
-        generatedDocumentInfo = GeneratedDocumentInfo.builder()
-            .url("TestUrl")
-            .binaryUrl("binaryUrl")
-            .hashToken("testHashToken")
-            .build();
-
-        List<FL401OrderTypeEnum> orderList = new ArrayList<>();
-
-        orderList.add(FL401OrderTypeEnum.occupationOrder);
-
-        orders = TypeOfApplicationOrders.builder()
-            .orderType(orderList)
-            .build();
-
-        linkToCA = LinkToCA.builder()
-            .linkToCaApplication(YesOrNo.Yes)
-            .caApplicationNumber("123")
-            .build();
-
-        PartyDetails applicant = PartyDetails.builder().representativeFirstName("Abc")
-            .representativeLastName("Xyz")
-            .gender(Gender.male)
-            .email("abc@xyz.com")
-            .canYouProvideEmailAddress(YesOrNo.Yes)
-            .phoneNumber("1234567890")
-            .isEmailAddressConfidential(YesOrNo.Yes)
-            .isPhoneNumberConfidential(YesOrNo.Yes)
-            .solicitorOrg(Organisation.builder().organisationID("ABC").organisationName("XYZ").build())
-            .solicitorAddress(Address.builder().addressLine1("ABC").postCode("AB1 2MN").build())
-            .doTheyHaveLegalRepresentation(YesNoDontKnow.yes)
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .typeOfApplicationOrders(orders)
-            .typeOfApplicationLinkToCA(linkToCA)
-            .draftOrderDoc(Document.builder()
-                               .documentUrl(generatedDocumentInfo.getUrl())
-                               .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                               .documentHash(generatedDocumentInfo.getHashToken())
-                               .documentFileName("FL401-Final.docx")
-                               .build())
-            .applicantsFL401(applicant)
-            .caseTypeOfApplication(PrlAppsConstants.FL401_CASE_TYPE)
-            .state(State.AWAITING_FL401_SUBMISSION_TO_HMCTS)
-            .welshLanguageRequirementApplication(LanguagePreference.english)
-            .languageRequirementApplicationNeedWelsh(YesOrNo.Yes)
-            .submitCountyCourtSelection(dynamicList)
-            .build();
-
-        CallbackResponse callbackResponse = CallbackResponse.builder()
-            .data(CaseData.builder()
-                      .draftOrderDoc(Document.builder()
-                                         .documentUrl(generatedDocumentInfo.getUrl())
-                                         .documentBinaryUrl(generatedDocumentInfo.getBinaryUrl())
-                                         .documentHash(generatedDocumentInfo.getHashToken())
-                                         .documentFileName("FL401-Final.docx")
-                                         .build())
-                      .state(State.AWAITING_SUBMISSION_TO_HMCTS)
-                      .build())
-            .build();
-
-        Map<String, Object> stringObjectMap = caseData.toMap(new ObjectMapper());
-
-        when(documentGenService.generateDocuments(Mockito.anyString(), Mockito.any(CaseData.class))).thenReturn(
-            fl401DocsMap);
-        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
-
-        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
-            .CallbackRequest.builder()
-            .caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder()
-                             .id(123L)
-                             .data(stringObjectMap)
-                             .build())
-            .build();
-
-        Court closestDomesticAbuseCourt = courtFinderService.getNearestFamilyCourt(
-            CaseUtils.getCaseData(callbackRequest.getCaseDetails(), objectMapper));
-        Optional<CourtEmailAddress> matchingEmailAddress = courtFinderService.getEmailAddress(closestDomesticAbuseCourt);
-
-        when(courtFinderService.getNearestFamilyCourt(CaseUtils.getCaseData(
-            callbackRequest.getCaseDetails(),
-            objectMapper
-        )))
-            .thenReturn(court);
-
-        UserDetails userDetails = UserDetails.builder()
-            .forename("test")
-            .surname("test")
-            .build();
-        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
-
-        fl401SubmitApplicationController.fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-
-        AboutToStartOrSubmitCallbackResponse response = fl401SubmitApplicationController
-            .fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
-
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_C8));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_FINAL));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_C8_WELSH));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_FINAL_WELSH));
-    }
-
-    @Test
     public void testCourtNameAndEmailAddressReturnedWhileFamilyEmailAddressReturned_WithBothOrders() throws Exception {
 
         generatedDocumentInfo = GeneratedDocumentInfo.builder()
@@ -1160,14 +380,16 @@ public class FL401SubmitApplicationControllerTest {
             .surname("test")
             .build();
         when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
+        when(fl401SubmitApplicationService.fl401GenerateDocumentSubmitApplication(
+            "test-auth",
+            callbackRequest,
+            caseData
+        )).thenReturn(stringObjectMap);
 
         AboutToStartOrSubmitCallbackResponse response = fl401SubmitApplicationController
             .fl401GenerateDocumentSubmitApplication(authToken, callbackRequest);
 
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_C8));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_FINAL));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_C8_WELSH));
-        assertTrue(response.getData().containsKey(DOCUMENT_FIELD_FINAL_WELSH));
+        assertNotNull(response.getData());
     }
 
     @Test
