+++ src/test/java/uk/gov/hmcts/reform/prl/services/LocationRefDataServiceTest.java
@@ -16,8 +16,8 @@ import uk.gov.hmcts.reform.prl.models.court.CourtDetails;
 import uk.gov.hmcts.reform.prl.models.court.CourtVenue;
 
 import java.util.List;
+import java.util.Optional;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
@@ -98,7 +98,7 @@ public class LocationRefDataServiceTest {
                                                      .courtEpimmsId("2")
                                                      .courtTypeId(FAMILY_COURT_TYPE_ID).build()))
                             .build());
-        String actual = locationRefDataService.getCourtDetailsFromEpimmsId("2", "test");
-        assertEquals("2-id-1-123-test-test", actual);
+        Optional<CourtVenue> courtVenue = locationRefDataService.getCourtDetailsFromEpimmsId("2", "test");
+        assertTrue(courtVenue.isPresent());
     }
 }
